+++ src/main/java/com/ddoganzip/entity/Order.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +36,7 @@ public class Order extends BaseEntity {
     private OrderStatus status = OrderStatus.CHECKING_STOCK;
 
     @Column(nullable = false)
-    private BigDecimal totalPrice;
+    private Integer totalPrice;
 
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL)
     private List<OrderItem> items = new ArrayList<>();
